creat modules and use ES6 syntax
