+++ src/index.js
@@ -1,22 +1,25 @@
 import './style.css';
 import {
   newGame,
-  postPlayer,
-  loadScores,
-  clearFields,
+  addPlayer,
 } from './module/module01.js';
+
+import {
+  updateScores,
+    blankInut,
+} from './module/module02.js';
 
 document.querySelector('.refresh-btn').addEventListener('click', (e) => {
   e.preventDefault();
-  loadScores(); 
+  updateScores(); 
 });
 
 document.getElementById('submit').addEventListener('click', (e) => {
   e.preventDefault();
   const name = document.querySelector('.typeName').value;
   const score = document.querySelector('.typeScore').value;
-  postPlayer(name, score); 
-  clearFields(); 
+  addPlayer(name, score); 
+  blankInut(); 
 });
 
 newGame();
